test checkTransmission
If we run this test often (+10 times) we see a 20% failure rate, although we didn't change anything between the runs.

Looking at the failiure, we see that since these are controls connected to a telemetry system sending live (simulated) values, we see that we never can be sure that the same value is retrieved in the diagnostic Info. So basically this is un-testable.
But we know we had the same problem with the TyrePressure Excercise.

Next we have to create the seam and stop accessing the real TelemetryClient.