adds our next intention by writing the test:
since the main problem of Alarm is that it directly depends on Sensor, we want to resolve this by injecting the Sensor object into Alarm.

And we want to be able to inject a static value into the sensor object what we want to inject into Alarm.

Like this Alarm becomes testable.