in Alarm we only achieved Dependency Injection of Sensor.
the changes from the IDE were to great, it changed the Sensor, which is not allowed in this kata.
So reverting that.

NEW REQUIREMENT:
We have a new requirement.
the current sensor and safetyRange are for "normal" tires, the new requirement is for wet tires.
The safetyRange for wet tires is [15;25]

and the Sensor has the specification
 OFFSET = 15
 samplePressure:
         double pressureTelemetryValue = 10 * basicRandomNumbersGenerator.nextDouble() * basicRandomNumbersGenerator.nextDouble();

Then we create the SensorNormalTire as an adapter to Sensor.