first version: we go step by step:
we see that the pressure is retrieved from the sensor directly in the check() method.
This means it's impossible for us to overwrite this.
So we exrtract this the call into a separate method.
Then we can overrride this method in our FakeAlarm
and return the specified pressure instead of returning the value directly from the sensor.

Bingo, the test is green.
Btw. is remains green also when running the test multiple times.
This was not the case originally, since the sensor returned random pressure values.