second version: we go step by step:
we "mock" the Sensor using Java's override in object initialisation.
but now we see again, that we need to pass the sensor into Alarm.

Actually this is quite good, because in the end we want to resolve the dependency on Sensor in Alarm (... see the new Sensor)

We trust the IDE to guide us in this refectoring:
* add paramter to Alarm's constructor
* add field to Alarm, manually the assignment "new Sensor", because now we have the sensor from the constructor

then we see that the IDE is telling us, that since we created the new constructor with arg, there is no default constructor any more.
we see that this hurts our test, because one of them fails.
we repair this, by using the IDE to create the default no-arg construtor.

Bingo, all tests are green.