we create the implementation of build.
here we see where we can use the new safetyRange object.
we simply add it to Alarm's constructor.

But this leads to making a vlot more changes until we get the test green.
we see that this was a too big change at one go.

but in the end Bingo, all tests are green.