since we have interfaces now, we can remove the duplicated code and adjust everything to use the interfaces.
Actually, using the interfaces we achive Inversion of Control = Dependency Inversion !

BUT we have a problem with the implementation of AlarmTest, because it needs to instantiate a safetyRange object in the setup.

We would like to parametrize the test, but since the test is using the mock runner we cannot add RunWith parametrized class.